refactor: modularize automation orchestrator
